+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- feat(ts): version-pinned git deps via #semver: + tag-push for npm/pnpm
+
 ## [9.0.0] - 2026-06-08
 
 ### Changed
@@ -552,6 +558,7 @@ at commit `9141ef54f5edac470d119a39285813299143898f`.
 
 > > > > > > > Stashed changes
 
+[Unreleased]: https://github.com/ggsuite/gg_one/compare/9.0.0...HEAD
 [9.0.0]: https://github.com/ggsuite/gg_one/compare/8.2.1...9.0.0
 [8.2.1]: https://github.com/ggsuite/gg_one/compare/8.2.0...8.2.1
 [8.2.0]: https://github.com/ggsuite/gg_one/compare/8.1.1...8.2.0